Inital Commit
The inital commit contains the base work I have done so far;

This includes 500k+ Tweets that have been harvested from twitter, alongside 3 files which combine them all under 'American/British collection' and 'Experimental'

It includes the R scripts that have been used to combine all the data into one cohesive structure with classification depending on party alignment.

It also includes the work so far with Python which to sumerize includes text classification for both american and british politics, as well as a breakdown of a inauguration speech which shows the party alignment per sentence.